requirements.txt
python-dotenv
openai
pdfminer.six
reportlab
pillow
imageio-ffmpeg
pydub
audioop-lts
requests
google-generativeai
